+++ threedi_models_and_simulations/data_models/simulation_data_models.py
@@ -107,11 +107,6 @@ class Laterals(SimulationElement):
 
 
 @dataclass
-class Substances(SimulationElement):
-    data: list = None
-
-
-@dataclass
 class DWF(SimulationElement):
     data: dict = None
 
@@ -203,7 +198,6 @@ class NewSimulation:
     structure_controls: StructureControls = None
     initial_conditions: InitialConditions = None
     laterals: Laterals = None
-    substances: Substances = None
     dwf: DWF = None
     breach: Breach = None
     precipitation: Precipitation = None